docs(docsite): teach the styling contract as the default, not an alternative
The docsite still narrated the attribute-and-stylesheet implementation as the
alternative to a Griffel-styled standard API. Both halves of that sentence are
now false: it is the only implementation, and there is no Griffel anywhere in
the package. Ten pages are truthed up against what ships.

`Icons/Headless API` becomes `Icons/Styling` and moves directly below the
catalog. A page whose premise was "headless vs standard" has no premise left,
but its content — the required stylesheet import, the cascade-layer obligation,
the `data-fui-icon*` mechanism table, the per-approach usage — is the only place
the docsite documents any of it, so the page is rewritten rather than retired.
The position change is the point: skipping the import is not an advanced topic,
it is the one mistake that silently breaks every icon in an application.

`Icons/Bundle Size` was a Griffel-vs-headless comparison whose closing advice
was "with webpack's default `style-loader` mode, prefer the Griffel variant" —
a recommendation to use something that no longer exists. It is rewritten against
the package's own `monosize` fixtures, which measure this repository rather than
an external application, and which say something the old page did not: the
`style-loader` regression is not a font-icon problem, SVG pays +20.1% too, and
the cost is fixed per application (6,631 B for one icon, 6,645 B for thirty-five)
rather than per icon. The external 35-icon table is kept as an approach ranking
with its provenance stated, because its absolute sizes do not reproduce here.

`Icons/API Contract` gains the third breaking-change category the package
`README.md` already carries: a styling-mechanism change requires a major. The
two documents are separate copies of the same contract, and leaving one of them
sanctioning only icon removal and renaming would leave the docsite disagreeing
with the package about the release it is documenting.

The rest is repointing and de-hedging. `Rendering Approaches` no longer offers a
headless counterpart to each approach — all three are styled identically now —
and its Setup row stops claiming inline SVG needs none. `Build Transforms` drops
the `/headless/*` targets and records that `headless: true` is deprecated for
this package, matching the loader's own README. The `Target every icon from CSS`
recipe mixed four flavours to prove `[data-fui-icon]` spans them; two of the four
were the same modules under a deprecated spelling, so it now mixes the two that
actually differ (`<svg>` and `<i>`) and gains the layering caveat that the
selector it teaches is subject to. Both File Type Icons pages get the same
treatment for `data-fui-filetype-icon`.

Every `/headless*` mention that survives is a deprecation-migration note.

Docsite build verified green: 26 entries, `icons-styling--docs` present,
`icons-headless-api--docs` gone with no dangling references to it.
