+++ packages/docsite/stories/FileTypeIcons/FileTypeIconsMigration.mdx
@@ -116,10 +116,11 @@ The v8 icon was rendered through the `<Icon>` component, which wrapped the glyph
 and applied container-oriented styles such as `overflow: hidden`. The v9 `FileTypeIcon` renders a
 **single `<img>`** with no wrapper, so styling applies directly to that element:
 
-- The default (Griffel) entry point and the opt-in `headless/styles.css` apply only what is
-  meaningful on a replaced element: `display: inline-block` and `object-fit: contain` (which keeps
-  non-square assets from stretching inside the square `width`/`height` box). `object-fit` is
-  **only** honored on the `<img>` itself — it has no effect on a parent element.
+- The shipped `@fluentui/react-icons-file-type/styles.css` applies only what is meaningful on a
+  replaced element: `display: inline-block` and `object-fit: contain` (which keeps non-square
+  assets from stretching inside the square `width`/`height` box). `object-fit` is **only** honored
+  on the `<img>` itself — it has no effect on a parent element. That stylesheet is a **required
+  import** in v9; there is no runtime style injection to fall back on.
 - The v8 `overflow: hidden` is intentionally **dropped**: `overflow` is a no-op on a replaced
   element like `<img>`, so it served no purpose.
 
@@ -145,6 +146,7 @@ on `object-fit`/`width`/`height` instead:
 | Icons registered into the global icon registry          | No global state; assets resolved per render                            |
 | Single density-resolved `src` (not SSR-safe)            | SSR-safe 1x `src` + `srcset`; browser picks the density                |
 | `<Icon>` wrapper with container styles (`overflow`)     | Single `<img>`, no wrapper; `object-fit` on the image (no `overflow`)  |
+| Styles injected at runtime                              | `import '@fluentui/react-icons-file-type/styles.css'` — required       |
 | `FileIconType` is a TS `enum` (with reverse mapping)    | `FileIconType` is a `const` object (no reverse mapping)                |
 
 ## APIs that were not ported
@@ -166,4 +168,4 @@ component instead:
 - **Hook:** `useFileTypeIconsContext`
 - **Values:** `FileIconType`, `DEFAULT_BASE_URL`, `FLUENT_CDN_BASE_URL`, `DEFAULT_ICON_SIZE`, `ICON_SIZES`
 - **Types:** `FileTypeIconProps`, `FileTypeIconsContextValue`, `FileTypeIconsProviderProps`, `FileIconTypeInput`, `FileTypeIconSize`, `ImageFileType`
-- **Entry points:** `@fluentui/react-icons-file-type` (default, Griffel-styled) and `@fluentui/react-icons-file-type/headless` (no Griffel runtime; opt-in `headless/styles.css`)
+- **Entry points:** `@fluentui/react-icons-file-type` and its required `@fluentui/react-icons-file-type/styles.css`. The `/headless` and `/headless/styles.css` subpaths are deprecated aliases of those two — same module, same file — and are removed in the next major.
